+++ batch3dfier/batch3dfierapp.py
@@ -308,6 +308,8 @@ def main():
     
     # If requires, copy the CSV output to postgres
     if cfg['out_table']:
+        config.create_heights_table(dbase, cfg['out_schema'], 
+                                    cfg['out_table'])
         with dbase.conn:
             with dbase.conn.cursor() as cur:
                 tbl = ".".join([cfg['out_schema'], cfg['out_table']])
@@ -315,13 +317,29 @@ def main():
                     # remove trailing commas from the CSV (until #58 is fixed in 3dfier)
                     command = "sed -i 's/,$//' " + p
                     call(command, shell=True)
+                    
                     with open(p, "r") as f_in:
                         # skip header
                         next(f_in)
                         cur.copy_from(f_in, tbl, sep=',')
+                        
                     # delete the CSV
                     command = "rm" + p
                     call(command, shell=True)
+        dbase.sendQuery(
+            sql.SQL("""CREATE INDEX IF NOT EXISTS {table}_id_idx 
+                            ON {schema}.{table} (id);
+                        """.format(schema=cfg['out_schema'],
+                                   table=cfg['out_table'])
+                    )
+        )
+        dbase.sendQuery(
+            sql.SQL("""COMMENT ON {schema}.{table} IS
+                    'Building heights generated with 3dfier.';
+                    """.format(schema=cfg['out_schema'],
+                                   table=cfg['out_table'])
+                    )
+        )
     else:
         pass